requirements.txt
fastapi
aiogram~=2.17.0
python-dotenv~=1.0.0
openai~=1.8.0
